Splitted up event_details.rs into multiple files
