+++ src-tauri/src/event_structures.rs
@@ -1,1 +1,3 @@
+pub mod deadline_details;
 pub mod event_details;
+pub mod event_type;
